requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Document Parsing
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.12

# LLM Integration
ollama==0.4.0
langchain==0.3.7
langchain-ollama==0.2.0
langchain-community==0.3.5

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
msal==1.31.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
reportlab==4.2.0

# Audio Interview (open-source)
faster-whisper==1.0.3
soundfile==0.12.1
numpy==1.26.4
